+++ malachite-base/tests/strings/random/random_fixed_length_strings.rs
@@ -8,7 +8,7 @@ use malachite_base::random::EXAMPLE_SEED;
 use malachite_base::strings::random::random_fixed_length_strings;
 
 fn random_fixed_length_strings_helper(
-    len: usize,
+    len: u64,
     expected_values: &[&str],
     expected_common_values: &[(&str, usize)],
     expected_median: (&str, Option<&str>),
